Add sidebar post processor
