+++ docs/documentation/litecommands/argument/argument/supported-types-adventure-extension.md
