+++ docs/documentation/litecommands/argument/argument/supported-types-bukkit-extension.md
