+++ docs/documentation/litecommands/argument/argument/supported-types-minestom-extension.md
